refactor: 에러 메시지 표시 기능 개선
- 에러 텍스트를 네트워크 상태에 따라 visible/gone으로 처리하는 것으로 변경
