+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="viewModel"
-            type="campus.tech.kakao.map.presentation.SearchViewModel" />
+            type="campus.tech.kakao.map.presentation.search.SearchViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -15,7 +15,7 @@
         android:layout_height="match_parent"
         android:paddingBottom="24dp"
         android:background="#FBFBFB"
-        tools:context=".presentation.SearchActivity">
+        tools:context=".presentation.search.SearchActivity">
 
         <androidx.appcompat.widget.LinearLayoutCompat
             android:id="@+id/navSearch"
